LayoutTests/http/tests/websocket/tests/hybi/invalid-continuation-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/invalid-continuation' failed: Received unexpected continuation frame.
Test whether WebSocket aborts the connection when it receives an unexpected continuation frame without any leading text or binary frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
